Refactor code
Add javadoc comments
Organize file into packages
Optimization for javafx
