--- src/main/java/Duke.java
@@ -1,152 +1,0 @@
-import duke.command.Command;
-import duke.DukeException;
-import duke.Parser;
-import duke.Storage;
-import duke.TaskList;
-import duke.Ui;
-
-import java.io.IOException;
-import java.text.ParseException;
-import java.time.format.DateTimeParseException;
-import java.util.Scanner;
-
-import javafx.application.Application;
-import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.control.ScrollPane;
-import javafx.scene.control.TextField;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.layout.AnchorPane;
-import javafx.scene.layout.Region;
-import javafx.scene.layout.VBox;
-import javafx.stage.Stage;
-
-public class Duke extends Application {
-    private ScrollPane scrollPane;
-    private VBox dialogContainer;
-    private TextField userInput;
-    private Button sendButton;
-    private Scene scene;
-    private Image user = new Image(this.getClass().getResourceAsStream("/images/user.png"));
-    private Image duke = new Image(this.getClass().getResourceAsStream("/images/duke.jpg"));
-    TaskList tasks = new TaskList();
-    Storage storage = new Storage();
-    boolean isFinished = false;
-
-
-    public static void main(String[] args) {
-        Scanner sc = new Scanner(System.in);
-
-
-
-        // TODO: 20/8/20 Improve runtime: keep an internal counter of task (for done.*)
-        Ui.initialMessage();
-//        while (!isFinished) {
-//            try {
-//                String echo = sc.nextLine();
-//                Command command = Parser.parseCommand(echo, tasks);
-//                command.execute(tasks);
-//                // System.out.println(tasks);
-//                isFinished = command.setIsFinished();
-//                storage.save(tasks);
-//            } catch (DukeException | ParseException | DateTimeParseException | IOException e) {
-//                System.out.println(e.getMessage());
-//            }
-//        }
-    }
-
-    @Override
-    public void start(Stage stage) throws Exception {
-        //Step 1. Setting up required components
-        // Add comment
-        //The container for the content of the chat to scroll.
-        scrollPane = new ScrollPane();
-        dialogContainer = new VBox();
-        scrollPane.setContent(dialogContainer);
-
-        userInput = new TextField();
-        sendButton = new Button("Send");
-
-        AnchorPane mainLayout = new AnchorPane();
-        mainLayout.getChildren().addAll(scrollPane, userInput, sendButton);
-
-        scene = new Scene(mainLayout);
-
-        stage.setScene(scene);
-        stage.show();
-
-        stage.setTitle("Duke");
-        stage.setResizable(false);
-        stage.setMinHeight(600.0);
-        stage.setMinWidth(400.0);
-
-        mainLayout.setPrefSize(400.0, 600.0);
-
-        scrollPane.setPrefSize(385, 535);
-        scrollPane.setHbarPolicy(ScrollPane.ScrollBarPolicy.NEVER);
-        scrollPane.setVbarPolicy(ScrollPane.ScrollBarPolicy.ALWAYS);
-
-        scrollPane.setVvalue(1.0);
-        scrollPane.setFitToWidth(true);
-
-        // You will need to import `javafx.scene.layout.Region` for this.
-        dialogContainer.setPrefHeight(Region.USE_COMPUTED_SIZE);
-
-        userInput.setPrefWidth(325.0);
-
-        sendButton.setPrefWidth(55.0);
-
-        AnchorPane.setTopAnchor(scrollPane, 1.0);
-
-        AnchorPane.setBottomAnchor(sendButton, 1.0);
-        AnchorPane.setRightAnchor(sendButton, 1.0);
-
-        AnchorPane.setLeftAnchor(userInput , 1.0);
-        AnchorPane.setBottomAnchor(userInput, 1.0);
-
-        sendButton.setOnMouseClicked((event) -> {
-            handleUserInput();
-        });
-
-        userInput.setOnAction((event) -> {
-            handleUserInput();
-        });
-
-        dialogContainer.heightProperty().addListener((observable) -> scrollPane.setVvalue(1.0));
-    }
-
-    private Label getDialogLabel(String text) {
-        // You will need to import `javafx.scene.control.Label`.
-        Label textToAdd = new Label(text);
-        textToAdd.setWrapText(true);
-
-        return textToAdd;
-    }
-
-    private void handleUserInput() {
-        String input = userInput.getText();
-        String response = getResponse(input);
-        dialogContainer.getChildren().addAll(
-                DialogBox.getUserDialog(input, user),
-                DialogBox.getDukeDialog(response, duke)
-        );
-        userInput.clear();
-    }
-
-    public String getResponse(String echo) {
-        Command command = null;
-        try {
-            command = Parser.parseCommand(echo, tasks);
-        } catch (DukeException | ParseException e) {
-            return e.getMessage();
-        }
-        String output = command.execute(tasks);
-        // System.out.println(tasks);
-        // storage.save(tasks);
-        return output;
-    }
-
-
-}
