--- src/main/java/duke/Task.java
@@ -1,35 +1,0 @@
-package duke;
-
-public class Task {
-    protected boolean isDone;
-    protected String description;
-
-    public Task(String description) {
-        this.isDone = false;
-        this.description = description;
-    }
-
-    public boolean isDone() {
-        return isDone;
-    }
-
-    public void setDone(boolean done) {
-        isDone = done;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public TaskType getType() {
-        return TaskType.NONE;
-    }
-
-    public String getDate() {
-        return "";
-    }
-
-    public String getDelimiter() {
-        return "";
-    }
-}
