+++ src/main/java/duke/command/DeadlineCommand.java
@@ -1,11 +1,11 @@
 package duke.command;
 
-import duke.Task;
-import duke.TaskList;
-import duke.Ui;
+import duke.logic.Ui;
+import duke.task.Task;
+import duke.task.TaskList;
 
 public class DeadlineCommand implements Command {
-    Task toBeAdded;
+    private Task toBeAdded;
 
     public DeadlineCommand(Task toBeAdded) {
         this.toBeAdded = toBeAdded;
