+++ src/main/java/duke/command/ToDoCommand.java
@@ -1,11 +1,11 @@
 package duke.command;
 
-import duke.Task;
-import duke.TaskList;
-import duke.Ui;
+import duke.logic.Ui;
+import duke.task.Task;
+import duke.task.TaskList;
 
 public class ToDoCommand implements Command {
-    Task toBeAdded;
+    private Task toBeAdded;
 
     public ToDoCommand(Task toBeAdded) {
         this.toBeAdded = toBeAdded;
