+++ src/main/java/duke/task/Event.java
@@ -1,21 +1,31 @@
-package duke;
+package duke.task;
 
 import java.text.ParseException;
-import java.text.SimpleDateFormat;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.util.Date;
 
 public class Event extends Task {
-    public LocalDateTime timeAt;
+    private LocalDateTime timeAt;
 
-    // TODO: 17/8/20 make a toString 
+    /**
+     * Constructs an Event object, which is a task with a time at which it is done.
+     * @param desc The description of the event.
+     * @param timeAt The time at which the event is done.
+     * @throws ParseException when the date time formatting is wrong.
+     */
+    // TODO: 17/8/20 make a toString
     public Event(String desc, String timeAt) throws ParseException {
         super(desc);
         DateTimeFormatter format = DateTimeFormatter.ofPattern("dd-MM-yyyy HHmm");
         this.timeAt = LocalDateTime.parse(timeAt, format);
     }
 
+    /**
+     * Returns a string representation of the event.
+     * The toString method for the Event object returns a string consisting of
+     * [E] indicating event, the description of the event, and the time of the event.
+     * @return A String representation of the Event object.
+     */
     @Override
     public String toString() {
         String sign = isDone ? "✓" : "✗";
