+++ src/main/java/duke/task/TaskList.java
@@ -1,18 +1,24 @@
-package duke;
+package duke.task;
 
 import java.util.ArrayList;
 
+import duke.common.Pair;
+
 public class TaskList {
-    public ArrayList<Task> thingsToDo;
+    private ArrayList<Task> thingsToDo;
 
     public TaskList() {
         this.thingsToDo = new ArrayList<>();
     }
 
-    // TODO: 17/8/20 print type, done
-    public Pair<String, Integer> add(Task detail) {
-        this.thingsToDo.add(detail);
-        return new Pair<>(detail.toString() + "\n", thingsToDo.size());
+    /**
+     * Adds a task to the list.
+     * @param task the task to be added
+     * @return a Pair, containing s String representation of the Task and the index number of the Task.
+     */
+    public Pair<String, Integer> add(Task task) {
+        this.thingsToDo.add(task);
+        return new Pair<>(task.toString() + "\n", thingsToDo.size());
         // TODO: 26/8/20   change this to length
     }
 
@@ -20,7 +26,11 @@ public class TaskList {
         return thingsToDo;
     }
 
-    public String printTodoList() {
+    /**
+     * Formats the list for printing purposes.
+     * @return a String representation of the list.
+     */
+    public String printTaskList() {
         String message = "";
         for (int i = 0; i < thingsToDo.size(); i++) {
             message = message
@@ -54,6 +64,10 @@ public class TaskList {
         return "Task{" + "thingsToDo=" + thingsToDo + '}';
     }
 
+    /**
+     * Formats the list to a String for saving purposes.
+     * @return a String representation of the list.
+     */
     public String toSave() {
         String res = "";
         for (Task task : thingsToDo) {
