+++ src/main/java/duke/task/TaskType.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.task;
 
 public enum TaskType {
     TODO {
