+++ text-ui-test/ACTUAL.TXT
@@ -1,5 +1,5 @@
     ____________________________________________________________
-    Hello! I'm Duke the Bad Dragon.
+    Hello! I'm duke.Duke the Bad Dragon.
     What can I do for you?
     ____________________________________________________________
 bruh you don't have spaces
